% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.enet.R
\name{perf}
\alias{perf}
\title{cross-validation function for elastic net panel}
\usage{
perf(x, ...)
}
\arguments{
\item{X}{nxp matrix - training dataset}

\item{Y}{categorical variables}

\item{alpha}{= 1 (lasso), alpha = 0 (ridge), 0 < alpha < 1 (elastic net penalty)}

\item{family}{"binomial" or "multinomial"}

\item{lambda}{= strength of elastic net penalty}

\item{M}{= # of folds}

\item{iter}{= number of times to repeat cross-valiation}

\item{threads}{- number of cpus (each cross-valiation scheme performed on a separate node)}

\item{progressBar}{- show progressbar (TRUE/FALE)}
}
\description{
Estimate test error of elastic net panel
}
